+++ src/components/Common/SsafyIcon/index.tsx
@@ -7,3 +7,4 @@ const SsafyIcon = {
 };
 
 export default SsafyIcon;
+export * from './Track';
