+++ src/components/UserRegisterForm/Fields/IsMember.tsx
@@ -1,8 +1,7 @@
 import { css } from '@emotion/react';
 import { isBoolean } from 'is-what';
 
-import { Button, SsafyIcon } from '~/components/Common';
-import { TrackSize } from '~/components/Common/SsafyIcon/Track';
+import { Button, SsafyIcon, TrackSize } from '~/components/Common';
 import { useUpdateMyInfoFormContext } from '~/services/member';
 import { flex } from '~/styles/utils';
